refactor: restructure acsl module
