+++ src/ebbflow/__init__.py
@@ -1,8 +1,8 @@
 from ebbflow.base_mechanistic_model import BaseMechanisticModel
 from ebbflow.acsl.decorators import INITIAL, DYNAMIC, DERIVATIVE, DISCRETE, TERMINAL, SORT
-from ebbflow.acsl.visitors.constant_collector import ConstantCollector
-from ebbflow.acsl.visitors.statevar_collector import StatevarCollector
-from ebbflow.acsl.preprocessing.acsl_sort import AcslSort
+from ebbflow.acsl.build.ast_visitors.constant_collector import ConstantCollector
+from ebbflow.acsl.build.ast_visitors.statevar_collector import StatevarCollector
+from ebbflow.acsl.build.sort.acsl_sort import AcslSort
 
 __all__ = [
     'BaseMechanisticModel',
